fix: home button readability and refactor course panel into dashboard
