--- src/component/course-panel/index.js
@@ -1,1 +1,0 @@
-export { default } from './course-panel.component'
