+++ src/component/dashboard/dashboard.reducer.js
@@ -1,4 +1,4 @@
-import { ADD_COURSE, REMOVE_COURSE, MODIFY_COURSE, INIT_COURSE_LIST } from './course-panel.const'
+import { ADD_COURSE, REMOVE_COURSE, MODIFY_COURSE, INIT_COURSE_LIST } from './dashboard.const'
 
 
 export const courseListReducer = (state = [], action) => {
